Background Sound Menu
I added a temporary background sound to the menu. When the player is on
the menu, the sound starts. The asset "Audio Menu Background" is
attached to the camera and set the position (0,0,0). If the position of
the asset "Audio Menu Background" is changed, the volume of the sound
will change though. You can also adjust the volume on the asset
inspector.