HUD 70% (LIFE/KILLS/WEAPON/BULLETS) DONE
IMPORTANT : Play the game with Maximize on Game to see the HUD

The HUD canvas display the current life of the player, the number of his
kills, the selected weapon, the time spend in the game and the number of
used bullets.

I modified something in PAUL script (WeaponSelector) in order to manage
the name of the selected weapon.
When the life is less than 50, it is displayed in red.

I let a empty text at the right in order to write the mission goal in
there.

To DO : Design a better HUD (add Textures, images)