Elevator enabled
Some buttons appear when you are in the elevator, they move you to
another level when you click on it with the mouse. I modified the build
settings and added the floor1 as the first floor of our building.
Note : the mouse has an issue and it is currently difficult to press on
the button, so use the pause menu to move the cursor.